Document installing Helm's Tiller
The installation of Helm is largely based on the work in
https://github.com/jonbcampos/kubernetes-series
and described in
https://medium.com/google-cloud/installing-helm-in-google-kubernetes-engine-7f07f43c536e
https://medium.com/google-cloud/install-secure-helm-in-gke-254d520061f7

This method sets up TLS between the Helm client and Tiller, so there's
some security there. I'm not sure I like how the keys are copied into
~/.helm, so I think that could be improved.

This method doesn't fully set up RBAC either (?) so that could be
improved.