Enable autodiscovery of topics
If kafkaefd salmock produce isn't given an explicit list of topics, it
now detects all SAL topics based on schemas in the schema registry.

First it lists all subjects in the registry, then downloads them and
finds which ones correspond to SAL schemas (based on the heuristic of
have a sal_topic_type key, but this could be changed to look for the
namespace instead.